BotContext interface changes
